+++ crates/nostr-indexeddb/src/lib.rs
@@ -462,8 +462,8 @@ impl_nostr_database!({
 
     #[inline]
     #[tracing::instrument(skip_all, level = "trace")]
-    async fn query(&self, filters: Vec<Filter>, order: Order) -> Result<Vec<Event>, DatabaseError> {
-        Ok(self.helper.query(filters, order).await)
+    async fn query(&self, filters: Vec<Filter>) -> Result<Vec<Event>, DatabaseError> {
+        Ok(self.helper.query(filters).await)
     }
 
     async fn negentropy_items(