+++ CHANGELOG.md
@@ -130,9 +130,12 @@
 * database: remove `NostrDatabase::bulk_import` ([Yuki Kishimoto])
 * database: remove `DatabaseError::NotFound` variant ([Yuki Kishimoto])
 * database: remove `DatabaseError::Nostr` variant ([Yuki Kishimoto])
+* database: remove `Order` enum ([Yuki Kishimoto])
+* database: remove `order` arg from `NostrDatabase::query` ([Yuki Kishimoto])
 * pool: remove high latency log ([Yuki Kishimoto])
 * zapper: remove `Err` from `NostrZapper` and unnecessary variants from `ZapperError` ([Yuki Kishimoto])
 * js(nostr): remove `Keys::vanity` ([Yuki Kishimoto])
+* cli: remove `reverse` flag from `query` command
 
 ## [v0.34.0]